+++ src/http/request.hpp
@@ -22,7 +22,7 @@ namespace http {
             inline const std::string& getPathStr() const { return pathStr; };
             inline const std::string& getBody() const { return body; };
             inline const std::string& getVersion() const { return httpVersionStr; };
-            inline int getCompressMethod() const { return compressMethod; };
+            int getCompressMethod(const std::string& MIME) const;
 
             void rewriteRawPath(const std::string& newPath);
 
@@ -62,7 +62,7 @@ namespace http {
             std::string httpVersionStr;
 
             std::string body;
-            int compressMethod = NO_COMPRESS;
+            int compressMethods = NO_COMPRESS;
     };
 
 }